Manual

HChebInterp.HAdaptError — Type

HAdaptError(; n=10)

Estimate the error of the interpolant by dividing the panel into two, computing interpolants on the subpanels, and computing the maximum error between interpolants at n*p equispaced points, where p is the number of points used to compute each interpolant.

HChebInterp.SpectralError — Type

SpectralError(; n=3)

Estimate the error of the interpolant by as the sum of the norm of the last n Chebyshev coefficients. Use abs to compute the norm of each coefficient.

HChebInterp.hchebinterp — Method

hchebinterp(f, a, b, [criterion=SpectralError()]; order=15, atol=0, rtol=0, norm=norm, maxevals=typemax(Int), initdiv=1)

Return a piecewise polynomial interpolant of f on the interval $[a,b]$ of degree order that is pointwise accurate to the requested tolerances. Uses criterion::AbstractAdaptCriterion to estimate the interpolant error for h-adaptation. If HAdaptError() is used as the criterion, it may be appropriate to reduce the order to 4 to avoid unnecessary function evaluations.

HChebInterp.AbstractAdaptCriterion — Type

AbstractAdaptCriterion

Abstract supertype for error criteria for adaptive refinement.